+++ src/main/java/com/devsuperior/DesafioORM/entities/Bloco.java
@@ -13,7 +13,7 @@ public class Bloco {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column (columnDefinition = "TIMESTAMP WITHOUT TIME ZONE")
     private Instant inicio;
@@ -28,17 +28,17 @@ public class Bloco {
     public Bloco (){
     }
 
-    public Bloco(Long id, Instant inicio, Instant fim) {
+    public Bloco(Integer id, Instant inicio, Instant fim) {
         this.id = id;
         this.inicio = inicio;
         this.fim = fim;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
